% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporaldynamics-package.R
\docType{package}
\name{spatiotemporaldynamics-package}
\alias{spatiotemporaldynamics}
\alias{spatiotemporaldynamics-package}
\title{spatiotemporaldynamics: Document data visualisation and GAM fitting for chickpea Ascochyta blight}
\description{
Research compendium documenting data visualisation and generalised
            additive model (GAM) fitting for the spatiotemporal dynamics of
            chickpea Ascochyta blight spread.
}
\author{
\strong{Maintainer}: Ihanul Khaliq \email{Ihsan.Khaliq@usq.edu.au}

}
\keyword{internal}
