% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatiotemporaldynamics-package.R
\docType{package}
\name{spatiotemporaldynamics-package}
\alias{spatiotemporaldynamics}
\alias{spatiotemporaldynamics-package}
\title{spatiotemporaldynamics: Spatiotemporal dynamics of chickpea ascochyta blight in chickpea}
\description{
Research compendium documenting data visualisation and generalised
            linear mixed models fitting to determine the spatiotemporal dynamics of
            chickpea Ascochyta blight in chickpea from primary infection foci in
            studies conducted in Queensland, AU during the 2020 winter 
            cropping season.
}
\author{
\strong{Maintainer}: Ihanul Khaliq \email{Ihsan.Khaliq@usq.edu.au}

Other contributors:
\itemize{
  \item Grains Research and Development Corporation (Provided support through funding GRDC Project
           USQ1903-003RTX) [funder, copyright holder]
  \item University of Southern Queensland [copyright holder]
}

}
\keyword{internal}
